Properly handle available databases at runtime and respect setup checks in command line as well
