+++ lib/private/util.php
@@ -497,7 +497,7 @@ class OC_Util {
 		}
 
 		$webServerRestart = false;
-		$setup = new OC\Setup($config);
+		$setup = new \OC\Setup($config, \OC::$server->getIniWrapper(), \OC::$server->getL10N('lib'), new \OC_Defaults());
 		$availableDatabases = $setup->getSupportedDatabases();
 		if (empty($availableDatabases)) {
 			$errors[] = array(
